# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FacilityTrack Aquatic Operations API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /work-orders:
    get:
      operationId: listWorkOrders
      tags: [work-orders]
      summary: List work orders
      parameters:
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: priority
          in: query
          schema:
            type: ["string", "null"]
        - name: assignedTo
          in: query
          schema:
            type: ["integer", "null"]
        - name: limit
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of work orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkOrder"
    post:
      operationId: createWorkOrder
      tags: [work-orders]
      summary: Create a work order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkOrderBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
  /work-orders/{id}:
    get:
      operationId: getWorkOrder
      tags: [work-orders]
      summary: Get a work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Work order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "404":
          description: Not found
    patch:
      operationId: updateWorkOrder
      tags: [work-orders]
      summary: Update a work order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkOrderBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
    delete:
      operationId: deleteWorkOrder
      tags: [work-orders]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── ASSETS ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    WorkOrder:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        facilityId:
          type: ["integer", "null"]
        poolId:
          type: ["integer", "null"]
        poolName:
          type: ["string", "null"]
        assetId:
          type: ["integer", "null"]
        assetName:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        status:
          type: string
          enum: [open, in_progress, completed, cancelled]
        assignedTo:
          type: ["integer", "null"]
        assignedToName:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: ["integer", "null"]
        createdByName:
          type: ["string", "null"]
        updatedAt:
          type: string
          format: date-time
        updatedBy:
          type: ["integer", "null"]
        updatedByName:
          type: ["string", "null"]
      required:
        - id
        - title
        - priority
        - status
        - createdAt
        - updatedAt
    CreateWorkOrderBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: ["string", "null"]
        facilityId:
          type: ["integer", "null"]
        poolId:
          type: ["integer", "null"]
        assetId:
          type: ["integer", "null"]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        assignedTo:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
      required:
        - title
        - priority
    UpdateWorkOrderBody:
      type: object
      properties:
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        priority:
          type: ["string", "null"]
        status:
          type: ["string", "null"]
        assignedTo:
          type: ["integer", "null"]
        assetId:
          type: ["integer", "null"]
        dueDate:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── ASSET ─────────────────────────────────────────────────────────────────
...